Time: 3 ms (45.22%), Space: 8.4 MB (92.25%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool isPalindrome(int x) {
-        long rev=0,y=(long)x;
+        long rev=0,y=x;
         if(x<0)
         return false;
             while(y!=0)
